+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - A notice suggesting to use the mouse to search the library,
 which is shown in the search bar when the app is running on a TV
+- Distinctive `User-Agent` to be shown in the PhotoPrism sessions list
 
 ### Changed
 - Improved quality of the viewed photos. Now the app loads corresponding previews